requirement.txt
APScheduler==3.6.0
beautifulsoup4==4.7.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cx-Oracle==7.2.2
filetype==1.0.5
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.2.5
MarkupSafe==1.1.1
Pillow==6.1.0
pycparser==2.19
pymssql==2.1.4
PyMuPDF==1.16.10
PyMySQL==0.9.3
pypinyin==0.35.4
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.5
suds-jurko==0.6
tornado==6.0.3
tzlocal==1.5.1
urllib3==1.25.3
Werkzeug==0.15.4
